refactor: move ValidationResult to errors namespace
