+++ vda5050_core/include/vda5050_core/order_utils/order_graph_validator.hpp
@@ -19,7 +19,7 @@
 #ifndef VDA5050_CORE__ORDER_UTILS__ORDER_GRAPH_VALIDATOR_HPP_
 #define VDA5050_CORE__ORDER_UTILS__ORDER_GRAPH_VALIDATOR_HPP_
 
-#include "vda5050_core/order_utils/validation_result.hpp"
+#include "vda5050_core/errors/validation_result.hpp"
 #include "vda5050_core/types/order.hpp"
 
 namespace vda5050_core {
@@ -31,16 +31,17 @@ namespace order_utils {
 ///
 /// \param order The order to be checked.
 ///
-/// \return A result struct of type ValidationResult
-ValidationResult is_valid_graph(const vda5050_core::types::Order& order);
+/// \return A result struct of type errors::ValidationResult
+errors::ValidationResult is_valid_graph(
+  const vda5050_core::types::Order& order);
 
 /// \brief Checks if order update is valid for order stitching
 ///
 /// \param base_order The base order.
 /// \param next_order the update order.
 ///
-/// \return A result struct of type ValidationResult
-ValidationResult is_valid_update(
+/// \return A result struct of type errors::ValidationResult
+errors::ValidationResult is_valid_update(
   const vda5050_core::types::Order& base_order,
   const vda5050_core::types::Order& next_order);
